baas(S1.8): address cpp + security review findings
SDK (game-facing, incl. wasm32 where long is 32-bit):
- JSON parser: recursion depth cap (stack-overflow DoS on malicious/MITM body),
  reject invalid/lone UTF-16 surrogates (was emitting garbage UTF-8).
- 64-bit wire integers: json Value + Session/Rank/Entry + submit() use long long,
  strtoll (was long/strtol -> silent truncation of user_id/score on wasm32/Win).
- Transport: cap response size (curl + emscripten) and null-check curl handles.
Backend:
- password.cc: sodium_memzero the hash buffer after use.
- lb_controller: try/catch parity in top()/me() (match submit/auth handlers).
- colony_scene: my_score_ long long + %lld formats.
Tests: sdk_client gains parser-hardening cases (deep nesting, bad surrogates,
2^53+1 no-truncation). All 19 suites green; ASan/UBSan clean; web SDK builds.

+++ sdk/cpp/include/gbaas/client.h
@@ -29,22 +29,22 @@ struct Config {
 };
 
 struct Session {
-    long        user_id = 0;
+    long long   user_id = 0;   // 64-bit: wasm32/Windows have 32-bit long
     std::string display_name;
     bool        is_guest = false;
 };
 
 struct Rank {
-    long value   = 0;
-    int  rank    = 0;
-    bool updated = false;
+    long long value   = 0;
+    int       rank    = 0;
+    bool      updated = false;
 };
 
 struct Entry {
     int         rank = 0;
-    long        user_id = 0;
+    long long   user_id = 0;
     std::string display_name;
-    long        value = 0;
+    long long   value = 0;
 };
 
 struct Board {
@@ -85,7 +85,7 @@ public:
 
     class Leaderboard {
     public:
-        void submit(long value, std::function<void(Result<Rank>)> cb);
+        void submit(long long value, std::function<void(Result<Rank>)> cb);
         void top(int limit, std::function<void(Result<Board>)> cb);
         void me(std::function<void(Result<Rank>)> cb);
     private:
